+++ src/widgets/layout/sidebar-layout/sidebar/menu/core-area.tsx
@@ -1,5 +1,6 @@
 import { usePathname } from "next/navigation";
 
+import { cn } from "@/shared/libs";
 import {
   Accordion,
   AccordionContent,
@@ -11,7 +12,7 @@ import {
 } from "@/shared/ui";
 
 import { SIDEBAR_GROUPS } from "../constants";
-import { CollapsibleMenu } from "./collapsible-menu";
+import { CollapsedMenu } from "./collapsed-menu";
 import { isItemActive } from "./helpers";
 import { MenuItem } from "./menu-item";
 import { Group } from "./types";
@@ -20,35 +21,35 @@ export function CoreAreaSidebarMenu() {
   const { open } = useSidebar();
   const pathname = usePathname() || "";
 
-  return open ? (
-    <SidebarMenu>
-      <AccordionMenu pathname={pathname} />
-    </SidebarMenu>
-  ) : (
-    <SidebarMenu className="p-2">
-      <CollapsibleMenu
-        items={Object.values(SIDEBAR_GROUPS).flatMap((group) =>
-          group.items.map((item, index) => ({
-            item,
-            group,
-            index,
-          }))
-        )}
-        getItemProps={({ item, group, index }) => {
-          const isGroupActive = group.items.some((item) =>
-            isItemActive(item.url, pathname)
-          );
+  return (
+    <SidebarMenu className={cn(!open && "p-2")}>
+      {open ? (
+        <AccordionMenu pathname={pathname} />
+      ) : (
+        <CollapsedMenu
+          items={Object.values(SIDEBAR_GROUPS).flatMap((group) =>
+            group.items.map((item, index) => ({
+              item,
+              group,
+              index,
+            }))
+          )}
+          getItemProps={({ item, group, index }) => {
+            const isGroupActive = group.items.some((item) =>
+              isItemActive(item.url, pathname)
+            );
 
-          return {
-            key: item.url,
-            label: item.title,
-            url: item.url,
-            Icon: item.icon,
-            hidden: !isGroupActive && index !== 0,
-            buttonClassName: isGroupActive ? "bg-sidebar-accent" : undefined,
-          };
-        }}
-      />
+            return {
+              key: item.url,
+              label: item.title,
+              url: item.url,
+              Icon: item.icon,
+              hidden: !isGroupActive && index !== 0,
+              buttonClassName: isGroupActive ? "bg-sidebar-accent" : undefined,
+            };
+          }}
+        />
+      )}
     </SidebarMenu>
   );
 }
